+++ CHANGELOG.md
@@ -18,6 +18,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
   * TODO: docs
 * New `craftsman example` or `craftsman new:example` command to create an example project with a prompted workflow
 * Added `.DS_Store` and `.env` to gitignore
+* Added Consumer test
+* Added `provider` to test fixture when adding a bus
 
 ### Updated